pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "synova-ai-project"
version = "0.1.0"
description = "Synova AI Project"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi>=0.95.0",
    "uvicorn[standard]>=0.22.0",
    "sqlalchemy>=2.0.21",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "pydantic>=1.10.0,<2.0.0",
    "python-multipart>=0.0.6"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "httpx>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0"
]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
